ExtendedLegacyOfRome: Extend "Restoration of Rome" event chain
Can now be done by any of the three contenders (HRE, Byzantium, Latin Empire),
or if none of those exists, by any roman, italian or grecoroman ruler.

HRE must be catholic (as their claim is based on papal authority), but the
rest may be of any cristian or pagan religion.